microbit_bsp
Module display

Driver a NxM LED matrix display

Can display 5x5 bitmaps from raw data or characters

Methods for scrolling text across LED matrix or displaying a bitmap for a duration

Modules

fonts

Bitmaps and fonts for the micro:bit

Structs

Bitmap

A bitmap with room for 8 bits used by Frame to create a compact frame buffer

Brightness

A brightness setting for the display.

Frame

An NxM frame that can be displayed on a LED matrix.

LedMatrix

Led matrix driver supporting arbitrary sized led matrixes.

Enums

AnimationEffect

An effect filter to apply for an animation

AnimationError

Errors produced when running animations